requirements.txt
Flask
flask-cors
gunicorn
openai
python-dotenv
gspread
google-auth